refactor: rename BlogView to ArticleView
